Read potmeter, fixes #73

+++ Raspberry-Pi/src/panel_control/go_to_angle.py
@@ -9,7 +9,7 @@ class GoToAngleThread(StoppableThread):
     # How close to try to get to the target
     accuracy = 0.1  # degrees
 
-    def __init__(self, angle: int, panel_controller):
+    def __init__(self, angle: float, panel_controller):
         super(GoToAngleThread, self).__init__()
         self.daemon = True
         self.target_angle = angle
